refactor structure and message
