+++ bot/__main__.py
@@ -3,16 +3,17 @@ import logging
 import os
 from datetime import datetime
 
-from aiogram import Bot, Dispatcher, Router
+from aiogram import Bot, Dispatcher
 from aiogram.filters import StateFilter
 from aiogram.filters.command import Command
 from aiogram.fsm.context import FSMContext
-from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import Message
 from dotenv import load_dotenv
 
 import database.config
 from bot.phrases_interpreter import read_file, read_placeholder_file
+from bot.routers.command_router import command_router
+from bot.states import States
 from database.dtos import NewSubscriptionDto, SubscriptionDto, UserDto
 from database.repositories import SubscriptionRepository, UserRepository
 
@@ -22,21 +23,14 @@ logging.basicConfig(level=logging.INFO)
 bot = Bot(token=os.getenv('TOKEN'))
 dispatcher = Dispatcher()
 
-router = Router()
-
 user_repository = UserRepository(session_maker=database.config.session_maker)
 subscription_repository = SubscriptionRepository(session_maker=database.config.session_maker)
 
 
-class States(StatesGroup):
-    registration_start = State()
-    change_name_start = State()
-    tmp = State()
-
-
 async def main():
+    dispatcher.include_router(command_router)
+    await bot.delete_webhook(drop_pending_updates=True)
     await dispatcher.start_polling(bot)
-    dispatcher.include_router(router)
 
 
 @dispatcher.message(StateFilter(States.registration_start))
@@ -78,47 +72,6 @@ async def send_welcome(message: Message, state: FSMContext) -> None:
         await state.set_state(States.registration_start)
     else:
         await message.answer(await read_placeholder_file('again_greeting.txt', message.from_user.id))
-
-
-@dispatcher.message(Command('help'))
-async def send_help(message: Message):
-    await bot.send_message(message.chat.id, read_file('help.txt'))
-
-
-@dispatcher.message(Command('rename'))
-async def send_rename(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('rename.txt'))
-    await state.set_state(States.change_name_start)
-
-
-@dispatcher.message(Command('enhance'))
-async def send_enhance(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
-
-
-@dispatcher.message(Command('recolor'))
-async def send_recolor(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
-
-
-@dispatcher.message(Command('change'))
-async def send_change(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
-
-
-@dispatcher.message(Command('generate'))
-async def send_generate(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
-
-
-@dispatcher.message(Command('effects'))
-async def send_effects(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
-
-
-@dispatcher.message(Command('prices'))
-async def send_prices(message: Message, state: FSMContext) -> None:
-    await message.answer(read_file('not_implemented.txt'))
 
 
 asyncio.run(main())
