+++ bot/phrases/help.txt
@@ -1,10 +1,10 @@
 Вот функционал, предоставляемый ботом:
 
-1) Улучшение качества изображений (/enhance)
-2) Перевод из черно-белого изображения в цветное (/recolor)
-3) Изменение изображение с помощью нейросети (/change)
-4) Генерация нового изображения (/generate)
-5) Добавление эффектов для изображения (/effects)
-6) Доступные тарифы и ограничения (/prices)
-7) Вызов справки по функционалу бота(/help)
-8) Изменить своё имя(/rename)
+1) /enhance Улучшение качества изображений
+2) /recolor Перевод из черно-белого изображения в цветное
+3) /change Изменение изображения с помощью нейросети
+4) /generate Генерация нового изображения
+5) /effects Добавление эффектов для изображения
+6) /prices Доступные тарифы и ограничения
+7) /help Вызов справки по функционалу бота
+8) /rename Изменить своё имя
