+++ bot/phrases/not_implemented.txt
@@ -1,3 +1,3 @@
-К сожалению данный функционал пока не доступен..
-Мы уже работаем над этим и обязательно сообщим, когда он станет доступен!
+К сожалению данный функционал пока не доступен...
+Мы уже работаем над этим и обязательно сообщим, когда он будет готов!
 Пожалуйста, воспользуйся другими командами
