+++ bot/states.py
@@ -1,0 +1,7 @@
+from aiogram.fsm.state import State, StatesGroup
+
+
+class States(StatesGroup):
+    registration_start = State()
+    change_name_start = State()
+    tmp = State()
